DI01 SPDVI : Sql Querys and Filters
I managed to filter the main results with SQL queries collecting the necessary columns inside the ComboBox and passing that data inside each query. I have several problems with the queries since I have used datagrid with DataTable and I don't really know how to filter the columns. I'm also trying to make a form when you click on a result, but I still don't know how to pass the data from the datagrid to the new form.

My idea is to create a new query within the base form that also collects the ID, in order to send it to the next form.

the aesthetic structure remains the same, functional for the moment.